Moved packing list components into packinglist directory
