+++ src/components/convention/packinglists/CostumeListSection.js
@@ -1,9 +1,8 @@
 import React, { Component } from 'react';
-import "../CosBuddy.css";
-import AppManager from '../../modules/AppManager';
-import CosPackingListCard from './CosPackingListCard'
+import "../../CosBuddy.css";
+import CostumePackingList from './CostumePackingList'
 
-export default class CostumePackingList extends Component {
+export default class CostumeListSection extends Component {
 
     // Set initial state
     state = {
@@ -47,7 +46,7 @@ export default class CostumePackingList extends Component {
                     {this.props.conCostumes.filter(conCostume =>
                             conCostume.userConventionId === this.props.convention.userConventionId)
                         .map(conCostume =>
-                        <CosPackingListCard key={conCostume.id} conCostume={conCostume} userConventionId={this.props.convention.userConventionId} {...this.props} />
+                        <CostumePackingList key={conCostume.id} conCostume={conCostume} userConventionId={this.props.convention.userConventionId} {...this.props} />
                     )}
 
                 </section>
